Tiny modification :fire:

+++ Daspro Semester 1/Jobsheet 7/ForKelipatan01.java
@@ -4,7 +4,7 @@ public class ForKelipatan01 {
     public static void main(String[] args) {
         Scanner Sc = new Scanner(System.in);
         int kelipatan, jumlah = 0, counter = 0;
-        int rataRata;
+        double rataRata;
         
 
         System.out.print("Masukkan bilangan kelipatan (1-9): ");
@@ -16,8 +16,8 @@ public class ForKelipatan01 {
                 counter++;
             }
         }
-        rataRata = jumlah / counter;
-        System.out.printf("Rata rata dari seluruh bagian kelipatan adalah %d\n",rataRata);
+        rataRata =  (double) jumlah / counter;
+        System.out.printf("Rata rata dari seluruh bagian kelipatan adalah %.2f\n",rataRata);
         System.out.printf("Banyaknya bilangan %d dari 1 sampai 50 adalah %d\n", kelipatan, counter);
         System.out.printf("Total bilangan kelipatan %d dari 1 sampai 50 adalah %d\n", kelipatan, jumlah);
